Use manually unsigned long long to get rid of the warning

+++ src/utils.c
@@ -18,12 +18,12 @@ typedef u32 size_t;
 typedef uint8_t u8;
 typedef uint16_t u16;
 typedef uint32_t u32;
-typedef uint64_t u64;
+typedef unsigned long long u64;
 
 typedef int8_t s8;
 typedef int16_t s16;
 typedef int32_t s32;
-typedef int64_t s64;
+typedef long long s64;
 
 typedef u32 bool;
 #endif
